Extract StarRating livewire component
